created a reusable moving maps background component
