+++ packages/components/package.json
@@ -79,14 +79,17 @@
     "svelte"
   ],
   "dependencies": {
+    "@allmaps/annotation": "workspace:^",
     "@allmaps/iiif-parser": "workspace:^",
     "@allmaps/project": "workspace:^",
     "@allmaps/stdlib": "workspace:^",
     "@allmaps/tailwind": "workspace:^",
     "@allmaps/transform": "workspace:^",
     "@allmaps/types": "workspace:^",
+    "@turf/rewind": "^7.2.0",
     "better-auth": "catalog:",
     "bits-ui": "^2.11.4",
+    "d3-geo": "^3.1.1",
     "flatbush": "^4.4.0",
     "fuse.js": "^7.1.0",
     "lodash-es": "^4.17.21",
@@ -101,6 +104,7 @@
     "@sveltejs/package": "^2.0.0",
     "@sveltejs/vite-plugin-svelte": "^7.0.0",
     "@tailwindcss/vite": "^4.0.0",
+    "@types/d3-geo": "^3.1.0",
     "@types/lodash-es": "^4.17.6",
     "eslint": "^9.18.0",
     "eslint-config-prettier": "^10.0.1",
